apps/recipe-url-importer/requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic-settings==2.7.0
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.3.1
readability-lxml==0.8.1
python-dateutil==2.9.0.post0
